wago: share native test fixtures across targets
Why:
- Linux CI could not compile tests whose helpers were missing behind unrelated platform tags
- TinyGo needs WABT-backed tests to remain compileable without invoking os/exec

What:
- centralize pure binary fixtures and WABT assembly helpers
- make WAT-dependent table tests return explicitly under TinyGo
- restore the missing cross-instance trap assertion helper

Proof:
- make test
- make test-guard
- WAGO_CORPUS_TIMEOUT=20s make test-corpus
- GOOS=linux GOARCH=amd64 go test -exec=true -count=1 ./...
- GOOS=linux GOARCH=arm64 go test -exec=true -count=1 ./...
- TinyGo linux/amd64 runtime and API test binaries compile

Docs:
- docs/tinygo.md updated
