+++ src/wago/multivalue_test.go
@@ -11,27 +11,6 @@ import (
 	"github.com/wago-org/wago/testutil/wasmtest"
 )
 
-func multiValueControlCallModule() []byte {
-	return wasmtest.Module(
-		wasmtest.Section(1, wasmtest.Vec(
-			wasmtest.FuncType(nil, []wasm.ValType{wasm.I32, wasm.I64}),
-			wasmtest.FuncType([]wasm.ValType{wasm.I32}, []wasm.ValType{wasm.I32, wasm.I64}),
-		)),
-		wasmtest.Section(3, wasmtest.Vec(
-			wasmtest.ULEB(0), // pair: () -> (i32, i64)
-			wasmtest.ULEB(1), // choose: (i32) -> (i32, i64)
-		)),
-		wasmtest.Section(7, wasmtest.Vec(
-			wasmtest.ExportEntry("pair", 0, 0),
-			wasmtest.ExportEntry("choose", 0, 1),
-		)),
-		wasmtest.Section(10, wasmtest.Vec(
-			wasmtest.Code([]byte{0x41, 0x12, 0x42, 0x13, 0x0b}),                                                 // i32.const 18; i64.const 19; end
-			wasmtest.Code([]byte{0x20, 0x00, 0x04, 0x00, 0x10, 0x00, 0x05, 0x41, 0x07, 0x42, 0x09, 0x0b, 0x0b}), // local.get 0; if type 0; call pair; else constants; end; end
-		)),
-	)
-}
-
 func multiValueBranchPayloadModule() []byte {
 	return wasmtest.Module(
 		wasmtest.Section(1, wasmtest.Vec(
